change app api to use tiktok api

+++ app/src/main/java/com/dinhlam/sharebox/extensions/UnitExtensions.kt
@@ -37,7 +37,7 @@ fun nowUTCTimeInMillis() = Calendar.getInstance().run {
     timeInMillis
 }
 
-fun Int.asViewCount(): String {
+fun Long.asViewCount(): String {
     val mil = this.toFloat() / 1_000_000
     val thous = this.toFloat() / 1_000
     return if (mil >= 1.0f) {
